+++ experiments/reinforcement_power_flow/congestion_helpers.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 import pandas as pd
 import pandapower as pp
